Use local db or dockertest for integration test with build tags
