+++ integration_test.go
@@ -1,16 +1,11 @@
-//go:build integration
-// +build integration
-
 package function
 
 import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"log"
 	"net/http"
 	"net/http/httptest"
-	"os"
 	"strings"
 	"testing"
 
@@ -18,75 +13,11 @@ import (
 	"github.com/operationspark/shorty/mongodb"
 	"github.com/operationspark/shorty/shorty"
 	"github.com/operationspark/shorty/testutil"
-	"github.com/ory/dockertest/v3"
-	"github.com/ory/dockertest/v3/docker"
-	"go.mongodb.org/mongo-driver/mongo"
-	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
-var dbClient *mongo.Client
-var dbName = "url-shortener-test"
-var urlCollName = "urls"
-
-func TestMain(m *testing.M) {
-	pool, err := dockertest.NewPool("")
-	if err != nil {
-		log.Fatalf("Could not connect to docker: %s", err)
-	}
-
-	// pull mongodb docker image for version 5.0
-	resource, err := pool.RunWithOptions(&dockertest.RunOptions{
-		Repository: "mongo",
-		Tag:        "5.0",
-		Env: []string{
-			// username and password for mongodb superuser
-			"MONGO_INITDB_ROOT_USERNAME=root",
-			"MONGO_INITDB_ROOT_PASSWORD=password",
-		},
-	}, func(config *docker.HostConfig) {
-		// set AutoRemove to true so that stopped container goes away by itself
-		config.AutoRemove = true
-		config.RestartPolicy = docker.RestartPolicy{
-			Name: "no",
-		}
-	})
-	if err != nil {
-		log.Fatalf("Could not start resource: %s", err)
-	}
-	// exponential backoff-retry, because the application in the container might not be ready to accept connections yet
-	err = pool.Retry(func() error {
-		var err error
-		dbClient, err = mongo.Connect(
-			context.TODO(),
-			options.Client().ApplyURI(
-				fmt.Sprintf("mongodb://root:password@localhost:%s", resource.GetPort("27017/tcp")),
-			),
-		)
-		if err != nil {
-			return err
-		}
-		return dbClient.Ping(context.TODO(), nil)
-	})
-
-	if err != nil {
-		log.Fatalf("Could not connect to docker: %s", err)
-	}
-
-	// run tests
-	code := m.Run()
-
-	// When you're done, kill and remove the container
-	if err = pool.Purge(resource); err != nil {
-		log.Fatalf("Could not purge resource: %s", err)
-	}
-
-	// disconnect mongodb client
-	if err = dbClient.Disconnect(context.TODO()); err != nil {
-		panic(err)
-	}
-
-	os.Exit(code)
-}
+// The MongoDB client for this tests are setup in
+// "integration-local-mongo_test.go" if the build tags do not include "integration",
+// or "integration-dockertest_test.go", if the build tags include "integration".
 
 func TestPOSTLinkIntegration(t *testing.T) {
 	t.Run("returns the Shorty by code", func(t *testing.T) {
